frontend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via
    #   django
    #   django-axes
babel==2.15.0
    # via -r requirements.in
backports-datetime-fromisoformat==2.0.1
    # via -r requirements.in
blessed==1.20.0
    # via curtsies
bpython==0.24
    # via
    #   -r requirements.in
    #   cdh-django-core
build==1.2.1
    # via pip-tools
cdh-django-core[all,core,federated-auth,files,recommended,rest,vue] @ git+https://github.com/CentreForDigitalHumanities/django-shared-core.git@v3.1.0
    # via -r requirements.in
certifi==2024.7.4
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
closure==20191111
    # via cdh-django-core
cryptography==42.0.8
    # via
    #   -r requirements.in
    #   cdh-django-core
    #   pyopenssl
    #   pysaml2
curtsies==0.4.2
    # via bpython
cwcwidth==0.1.9
    # via
    #   bpython
    #   curtsies
defusedxml==0.7.1
    # via
    #   djangosaml2
    #   pysaml2
deprecated==1.2.14
    # via cdh-django-core
django==4.2.14
    # via
    #   -r requirements.in
    #   cdh-django-core
    #   django-axes
    #   django-braces
    #   django-csp
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-modeltranslation
    #   django-simple-menu
    #   djangorestframework
    #   djangosaml2
django-axes==6.5.1
    # via -r requirements.in
django-braces==1.15.0
    # via
    #   -r requirements.in
    #   cdh-django-core
django-csp==3.8
    # via
    #   -r requirements.in
    #   cdh-django-core
django-debug-toolbar==4.4.6
    # via
    #   -r requirements.in
    #   cdh-django-core
    #   django-requests-debug-toolbar
django-extensions==3.2.3
    # via
    #   -r requirements.in
    #   cdh-django-core
django-filter==24.3
    # via -r requirements.in
django-impersonate==1.9.4
    # via cdh-django-core
django-modeltranslation==0.19.5
    # via cdh-django-core
django-requests-debug-toolbar==0.0.7
    # via -r requirements.in
django-simple-menu==2.1.3
    # via
    #   -r requirements.in
    #   cdh-django-core
djangorestframework==3.15.2
    # via cdh-django-core
djangosaml2==1.6.0
    # via cdh-django-core
elementpath==4.4.0
    # via xmlschema
greenlet==3.0.3
    # via bpython
idna==3.7
    # via requests
importlib-metadata==8.0.0
    # via build
lesscpy==0.15.1
    # via cdh-django-core
mysqlclient==2.2.4
    # via -r requirements.in
packaging==24.1
    # via build
pillow==10.4.0
    # via -r requirements.in
pip-tools==7.4.1
    # via -r requirements.in
ply==3.11
    # via lesscpy
pscript==0.7.7
    # via vbuild
pycparser==2.22
    # via cffi
pygments==2.18.0
    # via bpython
pyjwt==2.8.0
    # via
    #   -r requirements.in
    #   cdh-django-core
pyopenssl==24.1.0
    # via pysaml2
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pysaml2==7.5.0
    # via djangosaml2
pyscss==1.4.0
    # via cdh-django-core
python-dateutil==2.9.0.post0
    # via pysaml2
python-magic==0.4.27
    # via cdh-django-core
pytz==2024.1
    # via pysaml2
pyxdg==0.28
    # via bpython
requests==2.32.3
    # via
    #   -r requirements.in
    #   bpython
    #   cdh-django-core
    #   pysaml2
six==1.16.0
    # via
    #   blessed
    #   pyscss
    #   python-dateutil
sqlparse==0.5.1
    # via
    #   django
    #   django-debug-toolbar
tomli==2.0.1
    # via
    #   build
    #   pip-tools
typing-extensions==4.12.2
    # via
    #   asgiref
    #   django-modeltranslation
urllib3==2.2.2
    # via requests
vbuild==0.8.2
    # via cdh-django-core
wcwidth==0.2.13
    # via blessed
wheel==0.43.0
    # via pip-tools
wrapt==1.16.0
    # via deprecated
xmlschema==2.5.1
    # via pysaml2
zipp==3.19.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
